// src/models/api/eventDtoAggregateType.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * queen-api
 * API for Queen/Stromae
 * OpenAPI spec version: 5.6.1
 */

export type EventDtoAggregateType =
  (typeof EventDtoAggregateType)[keyof typeof EventDtoAggregateType]

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const EventDtoAggregateType = {
  QUESTIONNAIRE: 'QUESTIONNAIRE',
} as const