// src/models/api/eventDtoEventType.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * queen-api
 * API for Queen/Stromae
 * OpenAPI spec version: 5.6.1
 */

export type EventDtoEventType =
  (typeof EventDtoEventType)[keyof typeof EventDtoEventType]

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const EventDtoEventType = {
  MULTIMODE_MOVED: 'MULTIMODE_MOVED',
  QUESTIONNAIRE_INIT: 'QUESTIONNAIRE_INIT',
  QUESTIONNAIRE_LEAF_STATES_UPDATED: 'QUESTIONNAIRE_LEAF_STATES_UPDATED',
  QUESTIONNAIRE_UPDATED: 'QUESTIONNAIRE_UPDATED',
  QUESTIONNAIRE_COMPLETED: 'QUESTIONNAIRE_COMPLETED',
  QUESTIONNAIRE_VALIDATED: 'QUESTIONNAIRE_VALIDATED',
  QUESTIONNAIRE_SWITCH_CAPI: 'QUESTIONNAIRE_SWITCH_CAPI',
} as const