// src/models/api/eventPayloadDtoMode.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * queen-api
 * API for Queen/Stromae
 * OpenAPI spec version: 5.6.1
 */

export type EventPayloadDtoMode =
  (typeof EventPayloadDtoMode)[keyof typeof EventPayloadDtoMode]

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const EventPayloadDtoMode = {
  CATI: 'CATI',
  CAPI: 'CAPI',
  CAWI: 'CAWI',
  PAPI: 'PAPI',
} as const